[Chore][OPS] merge fragmented reduction op files into family modules
tileops/ops/reduction/ 16 files -> 7, matching the kernels-layer layout:

- cumulative.py <- cumulative_base, cumsum, cumprod
- softmax.py <- _softmax_base, softmax, log_softmax, logsumexp
- argreduce.py <- argmax, argmin
- logical_reduce.py <- all_op, any_op, count_nonzero
- vector_norm.py <- l1_norm, l2_norm, inf_norm

Class bodies unchanged; __all__ lists merged; importers retargeted;
manifest source.op pointers updated (reduction.yaml, scan.yaml).
Collected node set byte-identical; validator passes.
