+++ benchmarks/ops/bench_softmax.py
@@ -9,9 +9,7 @@ import torch
 import torch.nn.functional as F
 
 from benchmarks.benchmark_base import BenchmarkReport, ManifestBenchmark, workloads_to_params
-from tileops.ops.reduction.log_softmax import LogSoftmaxFwdOp
-from tileops.ops.reduction.logsumexp import LogSumExpFwdOp
-from tileops.ops.reduction.softmax import SoftmaxFwdOp
+from tileops.ops.reduction.softmax import LogSoftmaxFwdOp, LogSumExpFwdOp, SoftmaxFwdOp
 from workloads.reduction import (
     LogSoftmaxTest,
     LogSumExpTest,
