+++ tests/ops/test_argreduce.py
@@ -178,7 +178,7 @@ def _exact_compare(output: torch.Tensor, output_ref: torch.Tensor) -> None:
 
 @ArgreduceBasicFixture
 def test_argmax_op(m: int, n: int, dtype: torch.dtype) -> None:
-    from tileops.ops.reduction.argmax import ArgmaxFwdOp
+    from tileops.ops.reduction.argreduce import ArgmaxFwdOp
 
     test = ArgreduceTest(m, n, dtype, "argmax")
     op = ArgmaxFwdOp(dtype=dtype, dim=-1)
@@ -187,7 +187,7 @@ def test_argmax_op(m: int, n: int, dtype: torch.dtype) -> None:
 
 @ArgreduceNonContigFixture
 def test_argmax_non_contiguous(m: int, n: int, dtype: torch.dtype) -> None:
-    from tileops.ops.reduction.argmax import ArgmaxFwdOp
+    from tileops.ops.reduction.argreduce import ArgmaxFwdOp
 
     x_full = torch.randn(m, n * 2, dtype=dtype, device="cuda")
     x = x_full[:, :n]
@@ -200,7 +200,7 @@ def test_argmax_non_contiguous(m: int, n: int, dtype: torch.dtype) -> None:
 
 @Argreduce3DFixture
 def test_argmax_3d(batch: int, seq: int, hidden: int, dtype: torch.dtype) -> None:
-    from tileops.ops.reduction.argmax import ArgmaxFwdOp
+    from tileops.ops.reduction.argreduce import ArgmaxFwdOp
 
     x = torch.randn(batch, seq, hidden, dtype=dtype, device="cuda")
     op = ArgmaxFwdOp(dtype=dtype, dim=-1)
@@ -212,7 +212,7 @@ def test_argmax_3d(batch: int, seq: int, hidden: int, dtype: torch.dtype) -> Non
 
 @Argreduce4DFixture
 def test_argmax_4d(b0: int, b1: int, b2: int, n: int, dtype: torch.dtype) -> None:
-    from tileops.ops.reduction.argmax import ArgmaxFwdOp
+    from tileops.ops.reduction.argreduce import ArgmaxFwdOp
 
     x = torch.randn(b0, b1, b2, n, dtype=dtype, device="cuda")
     op = ArgmaxFwdOp(dtype=dtype, dim=-1)
@@ -224,7 +224,7 @@ def test_argmax_4d(b0: int, b1: int, b2: int, n: int, dtype: torch.dtype) -> Non
 
 @Argreduce1DFixture
 def test_argmax_1d(n: int, dtype: torch.dtype) -> None:
-    from tileops.ops.reduction.argmax import ArgmaxFwdOp
+    from tileops.ops.reduction.argreduce import ArgmaxFwdOp
 
     x = torch.randn(n, dtype=dtype, device="cuda")
     op = ArgmaxFwdOp(dtype=dtype, dim=-1)
@@ -237,7 +237,7 @@ def test_argmax_1d(n: int, dtype: torch.dtype) -> None:
 @Argreduce3DDim0Fixture
 def test_argmax_3d_dim0(batch: int, seq: int, hidden: int, dtype: torch.dtype) -> None:
     """Argmax along dim=0 on 3D tensors (outermost-dim reduction)."""
-    from tileops.ops.reduction.argmax import ArgmaxFwdOp
+    from tileops.ops.reduction.argreduce import ArgmaxFwdOp
 
     x = torch.randn(batch, seq, hidden, dtype=dtype, device="cuda")
     op = ArgmaxFwdOp(dtype=dtype, dim=0)
@@ -251,7 +251,7 @@ def test_argmax_3d_dim0(batch: int, seq: int, hidden: int, dtype: torch.dtype) -
 @Argreduce3DDim0Fixture
 def test_argmax_3d_dim0_keepdim(batch: int, seq: int, hidden: int, dtype: torch.dtype) -> None:
     """Argmax along dim=0 with keepdim=True on 3D tensors."""
-    from tileops.ops.reduction.argmax import ArgmaxFwdOp
+    from tileops.ops.reduction.argreduce import ArgmaxFwdOp
 
     x = torch.randn(batch, seq, hidden, dtype=dtype, device="cuda")
     op = ArgmaxFwdOp(dtype=dtype, dim=0, keepdim=True)
@@ -265,7 +265,7 @@ def test_argmax_3d_dim0_keepdim(batch: int, seq: int, hidden: int, dtype: torch.
 @Argreduce4DDim0Fixture
 def test_argmax_4d_dim0(b0: int, b1: int, b2: int, n: int, dtype: torch.dtype) -> None:
     """Argmax along dim=0 on 4D tensors (outermost-dim reduction, 3D+ regression)."""
-    from tileops.ops.reduction.argmax import ArgmaxFwdOp
+    from tileops.ops.reduction.argreduce import ArgmaxFwdOp
 
     x = torch.randn(b0, b1, b2, n, dtype=dtype, device="cuda")
     op = ArgmaxFwdOp(dtype=dtype, dim=0)
@@ -279,7 +279,7 @@ def test_argmax_4d_dim0(b0: int, b1: int, b2: int, n: int, dtype: torch.dtype) -
 @Argreduce4DDim0Fixture
 def test_argmax_4d_dim0_keepdim(b0: int, b1: int, b2: int, n: int, dtype: torch.dtype) -> None:
     """Argmax along dim=0 with keepdim=True on 4D tensors."""
-    from tileops.ops.reduction.argmax import ArgmaxFwdOp
+    from tileops.ops.reduction.argreduce import ArgmaxFwdOp
 
     x = torch.randn(b0, b1, b2, n, dtype=dtype, device="cuda")
     op = ArgmaxFwdOp(dtype=dtype, dim=0, keepdim=True)
@@ -293,7 +293,7 @@ def test_argmax_4d_dim0_keepdim(b0: int, b1: int, b2: int, n: int, dtype: torch.
 @SpecArgreduceFixture
 def test_argmax_spec_dim(shape: tuple, dim: int, keepdim: bool, dtype: torch.dtype) -> None:
     """Spec interface: ArgmaxFwdOp with dim + keepdim."""
-    from tileops.ops.reduction.argmax import ArgmaxFwdOp
+    from tileops.ops.reduction.argreduce import ArgmaxFwdOp
 
     x = torch.randn(*shape, dtype=dtype, device="cuda")
     op = ArgmaxFwdOp(dtype=dtype, dim=dim, keepdim=keepdim)
@@ -309,7 +309,7 @@ def test_argmax_spec_dim(shape: tuple, dim: int, keepdim: bool, dtype: torch.dty
 
 @ArgreduceBasicFixture
 def test_argmin_op(m: int, n: int, dtype: torch.dtype) -> None:
-    from tileops.ops.reduction.argmin import ArgminFwdOp
+    from tileops.ops.reduction.argreduce import ArgminFwdOp
 
     test = ArgreduceTest(m, n, dtype, "argmin")
     op = ArgminFwdOp(dtype=dtype, dim=-1)
@@ -318,7 +318,7 @@ def test_argmin_op(m: int, n: int, dtype: torch.dtype) -> None:
 
 @ArgreduceNonContigFixture
 def test_argmin_non_contiguous(m: int, n: int, dtype: torch.dtype) -> None:
-    from tileops.ops.reduction.argmin import ArgminFwdOp
+    from tileops.ops.reduction.argreduce import ArgminFwdOp
 
     x_full = torch.randn(m, n * 2, dtype=dtype, device="cuda")
     x = x_full[:, :n]
@@ -331,7 +331,7 @@ def test_argmin_non_contiguous(m: int, n: int, dtype: torch.dtype) -> None:
 
 @Argreduce3DFixture
 def test_argmin_3d(batch: int, seq: int, hidden: int, dtype: torch.dtype) -> None:
-    from tileops.ops.reduction.argmin import ArgminFwdOp
+    from tileops.ops.reduction.argreduce import ArgminFwdOp
 
     x = torch.randn(batch, seq, hidden, dtype=dtype, device="cuda")
     op = ArgminFwdOp(dtype=dtype, dim=-1)
@@ -343,7 +343,7 @@ def test_argmin_3d(batch: int, seq: int, hidden: int, dtype: torch.dtype) -> Non
 
 @Argreduce4DFixture
 def test_argmin_4d(b0: int, b1: int, b2: int, n: int, dtype: torch.dtype) -> None:
-    from tileops.ops.reduction.argmin import ArgminFwdOp
+    from tileops.ops.reduction.argreduce import ArgminFwdOp
 
     x = torch.randn(b0, b1, b2, n, dtype=dtype, device="cuda")
     op = ArgminFwdOp(dtype=dtype, dim=-1)
@@ -355,7 +355,7 @@ def test_argmin_4d(b0: int, b1: int, b2: int, n: int, dtype: torch.dtype) -> Non
 
 @Argreduce1DFixture
 def test_argmin_1d(n: int, dtype: torch.dtype) -> None:
-    from tileops.ops.reduction.argmin import ArgminFwdOp
+    from tileops.ops.reduction.argreduce import ArgminFwdOp
 
     x = torch.randn(n, dtype=dtype, device="cuda")
     op = ArgminFwdOp(dtype=dtype, dim=-1)
@@ -368,7 +368,7 @@ def test_argmin_1d(n: int, dtype: torch.dtype) -> None:
 @Argreduce3DDim0Fixture
 def test_argmin_3d_dim0(batch: int, seq: int, hidden: int, dtype: torch.dtype) -> None:
     """Argmin along dim=0 on 3D tensors (outermost-dim reduction)."""
-    from tileops.ops.reduction.argmin import ArgminFwdOp
+    from tileops.ops.reduction.argreduce import ArgminFwdOp
 
     x = torch.randn(batch, seq, hidden, dtype=dtype, device="cuda")
     op = ArgminFwdOp(dtype=dtype, dim=0)
@@ -382,7 +382,7 @@ def test_argmin_3d_dim0(batch: int, seq: int, hidden: int, dtype: torch.dtype) -
 @Argreduce3DDim0Fixture
 def test_argmin_3d_dim0_keepdim(batch: int, seq: int, hidden: int, dtype: torch.dtype) -> None:
     """Argmin along dim=0 with keepdim=True on 3D tensors."""
-    from tileops.ops.reduction.argmin import ArgminFwdOp
+    from tileops.ops.reduction.argreduce import ArgminFwdOp
 
     x = torch.randn(batch, seq, hidden, dtype=dtype, device="cuda")
     op = ArgminFwdOp(dtype=dtype, dim=0, keepdim=True)
@@ -396,7 +396,7 @@ def test_argmin_3d_dim0_keepdim(batch: int, seq: int, hidden: int, dtype: torch.
 @Argreduce4DDim0Fixture
 def test_argmin_4d_dim0(b0: int, b1: int, b2: int, n: int, dtype: torch.dtype) -> None:
     """Argmin along dim=0 on 4D tensors (outermost-dim reduction, 3D+ regression)."""
-    from tileops.ops.reduction.argmin import ArgminFwdOp
+    from tileops.ops.reduction.argreduce import ArgminFwdOp
 
     x = torch.randn(b0, b1, b2, n, dtype=dtype, device="cuda")
     op = ArgminFwdOp(dtype=dtype, dim=0)
@@ -410,7 +410,7 @@ def test_argmin_4d_dim0(b0: int, b1: int, b2: int, n: int, dtype: torch.dtype) -
 @Argreduce4DDim0Fixture
 def test_argmin_4d_dim0_keepdim(b0: int, b1: int, b2: int, n: int, dtype: torch.dtype) -> None:
     """Argmin along dim=0 with keepdim=True on 4D tensors."""
-    from tileops.ops.reduction.argmin import ArgminFwdOp
+    from tileops.ops.reduction.argreduce import ArgminFwdOp
 
     x = torch.randn(b0, b1, b2, n, dtype=dtype, device="cuda")
     op = ArgminFwdOp(dtype=dtype, dim=0, keepdim=True)
@@ -424,7 +424,7 @@ def test_argmin_4d_dim0_keepdim(b0: int, b1: int, b2: int, n: int, dtype: torch.
 @SpecArgreduceFixture
 def test_argmin_spec_dim(shape: tuple, dim: int, keepdim: bool, dtype: torch.dtype) -> None:
     """Spec interface: ArgminFwdOp with dim + keepdim."""
-    from tileops.ops.reduction.argmin import ArgminFwdOp
+    from tileops.ops.reduction.argreduce import ArgminFwdOp
 
     x = torch.randn(*shape, dtype=dtype, device="cuda")
     op = ArgminFwdOp(dtype=dtype, dim=dim, keepdim=keepdim)
@@ -493,7 +493,7 @@ class ArgreduceDimNoneFixture(FixtureBase):
 @ArgreduceDimNoneFixture
 def test_argmax_dim_none(shape: tuple, dtype: torch.dtype) -> None:
     """ArgmaxFwdOp(dim=None) matches torch.argmax(x); covers keepdim={False, True}."""
-    from tileops.ops.reduction.argmax import ArgmaxFwdOp
+    from tileops.ops.reduction.argreduce import ArgmaxFwdOp
 
     x = torch.randn(*shape, dtype=dtype, device="cuda")
     ref_flat = torch.argmax(x)
@@ -514,7 +514,7 @@ def test_argmax_dim_none(shape: tuple, dtype: torch.dtype) -> None:
 @ArgreduceDimNoneFixture
 def test_argmin_dim_none(shape: tuple, dtype: torch.dtype) -> None:
     """ArgminFwdOp(dim=None) matches torch.argmin(x); covers keepdim={False, True}."""
-    from tileops.ops.reduction.argmin import ArgminFwdOp
+    from tileops.ops.reduction.argreduce import ArgminFwdOp
 
     x = torch.randn(*shape, dtype=dtype, device="cuda")
     ref_flat = torch.argmin(x)
