+++ tests/ops/test_reduction_defaults.py
@@ -113,7 +113,7 @@ def test_var_mean_default_dim_full_reduction() -> None:
 
 @pytest.mark.smoke
 def test_all_default_dim_full_reduction() -> None:
-    from tileops.ops.reduction.all_op import AllFwdOp
+    from tileops.ops.reduction.logical_reduce import AllFwdOp
 
     x = _make_logical(_LOGICAL_SHAPE, torch.float16)
     op = AllFwdOp(dtype=torch.float16)
@@ -124,7 +124,7 @@ def test_all_default_dim_full_reduction() -> None:
 
 @pytest.mark.smoke
 def test_any_default_dim_full_reduction() -> None:
-    from tileops.ops.reduction.any_op import AnyFwdOp
+    from tileops.ops.reduction.logical_reduce import AnyFwdOp
 
     x = _make_logical(_LOGICAL_SHAPE, torch.float16)
     op = AnyFwdOp(dtype=torch.float16)
@@ -135,7 +135,7 @@ def test_any_default_dim_full_reduction() -> None:
 
 @pytest.mark.smoke
 def test_count_nonzero_default_dim_full_reduction() -> None:
-    from tileops.ops.reduction.count_nonzero import CountNonzeroFwdOp
+    from tileops.ops.reduction.logical_reduce import CountNonzeroFwdOp
 
     x = _make_logical(_LOGICAL_SHAPE, torch.float16)
     op = CountNonzeroFwdOp(dtype=torch.float16)
@@ -164,7 +164,7 @@ def test_prod_default_dim_last_axis() -> None:
 @pytest.mark.smoke
 @pytest.mark.parametrize("empty_dim", [[], ()])
 def test_all_empty_dim_noop(empty_dim) -> None:
-    from tileops.ops.reduction.all_op import AllFwdOp
+    from tileops.ops.reduction.logical_reduce import AllFwdOp
 
     x = _make_logical(_LOGICAL_SHAPE, torch.float16)
     op = AllFwdOp(dtype=torch.float16, dim=empty_dim)
@@ -177,7 +177,7 @@ def test_all_empty_dim_noop(empty_dim) -> None:
 @pytest.mark.smoke
 @pytest.mark.parametrize("empty_dim", [[], ()])
 def test_any_empty_dim_noop(empty_dim) -> None:
-    from tileops.ops.reduction.any_op import AnyFwdOp
+    from tileops.ops.reduction.logical_reduce import AnyFwdOp
 
     x = _make_logical(_LOGICAL_SHAPE, torch.float16)
     op = AnyFwdOp(dtype=torch.float16, dim=empty_dim)
@@ -216,9 +216,7 @@ def test_normalize_dim_full_returns_all() -> None:
 @pytest.mark.smoke
 def test_empty_dim_policy_class_attrs() -> None:
     """Per-op empty_dim_policy bindings."""
-    from tileops.ops.reduction.all_op import AllFwdOp
-    from tileops.ops.reduction.any_op import AnyFwdOp
-    from tileops.ops.reduction.count_nonzero import CountNonzeroFwdOp
+    from tileops.ops.reduction.logical_reduce import AllFwdOp, AnyFwdOp, CountNonzeroFwdOp
     from tileops.ops.reduction.reduce import (
         AmaxFwdOp,
         AminFwdOp,
@@ -249,7 +247,7 @@ def test_empty_dim_policy_class_attrs() -> None:
 @pytest.mark.smoke
 def test_all_empty_dim_noop_rejects_cpu_tensor() -> None:
     """dim=[] must still validate device; non-CUDA input must raise."""
-    from tileops.ops.reduction.all_op import AllFwdOp
+    from tileops.ops.reduction.logical_reduce import AllFwdOp
 
     x = (torch.randint(-1, 2, _LOGICAL_SHAPE)).to(torch.float16)  # cpu
     op = AllFwdOp(dtype=torch.float16, dim=[])
@@ -259,7 +257,7 @@ def test_all_empty_dim_noop_rejects_cpu_tensor() -> None:
 
 @pytest.mark.smoke
 def test_any_empty_dim_noop_rejects_cpu_tensor() -> None:
-    from tileops.ops.reduction.any_op import AnyFwdOp
+    from tileops.ops.reduction.logical_reduce import AnyFwdOp
 
     x = (torch.randint(-1, 2, _LOGICAL_SHAPE)).to(torch.float16)  # cpu
     op = AnyFwdOp(dtype=torch.float16, dim=[])
@@ -270,7 +268,7 @@ def test_any_empty_dim_noop_rejects_cpu_tensor() -> None:
 @pytest.mark.smoke
 def test_all_empty_dim_noop_rejects_wrong_dtype() -> None:
     """dim=[] must still validate dtype against the op's declared dtype."""
-    from tileops.ops.reduction.all_op import AllFwdOp
+    from tileops.ops.reduction.logical_reduce import AllFwdOp
 
     x = _make_logical(_LOGICAL_SHAPE, torch.float32)  # cuda, fp32
     op = AllFwdOp(dtype=torch.float16, dim=[])
@@ -280,7 +278,7 @@ def test_all_empty_dim_noop_rejects_wrong_dtype() -> None:
 
 @pytest.mark.smoke
 def test_any_empty_dim_noop_rejects_wrong_dtype() -> None:
-    from tileops.ops.reduction.any_op import AnyFwdOp
+    from tileops.ops.reduction.logical_reduce import AnyFwdOp
 
     x = _make_logical(_LOGICAL_SHAPE, torch.float32)
     op = AnyFwdOp(dtype=torch.float16, dim=[])
@@ -293,7 +291,7 @@ def test_all_empty_dim_noop_binds_roofline() -> None:
     """eval_roofline() must succeed after a dim=[] noop forward and
     report non-zero data-movement (the noop still reads the input and
     writes an equal-shape cast result)."""
-    from tileops.ops.reduction.all_op import AllFwdOp
+    from tileops.ops.reduction.logical_reduce import AllFwdOp
 
     x = _make_logical(_LOGICAL_SHAPE, torch.float16)
     op = AllFwdOp(dtype=torch.float16, dim=[])
@@ -316,7 +314,7 @@ def test_all_empty_dim_noop_binds_roofline() -> None:
 
 @pytest.mark.smoke
 def test_any_empty_dim_noop_binds_roofline() -> None:
-    from tileops.ops.reduction.any_op import AnyFwdOp
+    from tileops.ops.reduction.logical_reduce import AnyFwdOp
 
     x = _make_logical(_LOGICAL_SHAPE, torch.float16)
     op = AnyFwdOp(dtype=torch.float16, dim=[])
