+++ tests/ops/test_softmax.py
@@ -18,9 +18,7 @@ import torch
 import torch.nn.functional as F
 
 from tests.test_base import FixtureBase, TestBase
-from tileops.ops.reduction.log_softmax import LogSoftmaxFwdOp
-from tileops.ops.reduction.logsumexp import LogSumExpFwdOp
-from tileops.ops.reduction.softmax import SoftmaxFwdOp
+from tileops.ops.reduction.softmax import LogSoftmaxFwdOp, LogSumExpFwdOp, SoftmaxFwdOp
 from workloads.reduction import LogSoftmaxTest as _LogSoftmaxTestWorkload
 from workloads.reduction import LogSumExpTest as _LogSumExpTestWorkload
 from workloads.reduction import SoftmaxTest as _SoftmaxTestWorkload
