+++ tileops/ops/reduction/cumulative.py
@@ -1,13 +1,4 @@
-"""CumulativeOp base class for scan operators (cumsum, cumprod).
-
-Implements manifest-driven dynamic shape binding: ctor binds only
-semantic/config params (`dim`, optional committed `N` for strict
-compatibility); `M` and the effective `N` are derived at forward time, kernels
-are built lazily and cached by `(M, N, dtype, device)`.
-
-Forward pipeline: validate -> movedim(dim → -1) -> reshape (M, N) -> kernel
-(handles alignment via masked loads) -> trim -> reshape -> movedim(-1 → dim).
-"""
+"""Cumulative scan operators (cumsum, cumprod)."""
 
 from abc import abstractmethod
 from typing import Dict, Optional, Tuple
@@ -20,7 +11,7 @@ from tileops.kernels.reduction.cumulative import CumulativeKernel
 
 from ..op_base import Op
 
-__all__ = ["CumulativeOp"]
+__all__ = ["CumprodFwdOp", "CumsumFwdOp", "CumulativeOp"]
 
 
 class CumulativeOp(Op):
@@ -143,3 +134,57 @@ class CumulativeOp(Op):
         if dim_norm != ndim - 1:
             y = y.movedim(-1, dim_norm)
         return y
+
+
+
+class CumsumFwdOp(CumulativeOp):
+    """Cumulative sum operator: ``y = cumsum(x, dim)``.
+
+    Output has the same shape and dtype as ``x``. Alignment padding is
+    handled inside the kernel via masked loads.
+
+    Args:
+        dtype: Optional data type (float32, float16, or bfloat16).
+            Preferred API infers it from ``x``.
+        dim: Reduction axis (default -1). Negative values are normalized
+            at forward time.
+        kernel_map: Optional override for kernel dispatch.
+        tune: Whether to autotune (default False).
+
+    Example:
+        >>> op = CumsumFwdOp()
+        >>> x = torch.randn(1024, 4096, dtype=torch.float16, device="cuda")
+        >>> y = op(x)  # shape: (1024, 4096)
+    """
+
+    _op_kind = "sum"
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        return self._run(x)
+
+
+
+class CumprodFwdOp(CumulativeOp):
+    """Cumulative product operator: ``y = cumprod(x, dim)``.
+
+    Output has the same shape and dtype as ``x``. Alignment padding is
+    handled inside the kernel via masked loads.
+
+    Args:
+        dtype: Optional data type (float32, float16, or bfloat16).
+            Preferred API infers it from ``x``.
+        dim: Reduction axis (default -1). Negative values are normalized
+            at forward time.
+        kernel_map: Optional override for kernel dispatch.
+        tune: Whether to autotune (default False).
+
+    Example:
+        >>> op = CumprodFwdOp()
+        >>> x = torch.randn(1024, 4096, dtype=torch.float16, device="cuda") * 0.01 + 0.99
+        >>> y = op(x)  # shape: (1024, 4096)
+    """
+
+    _op_kind = "prod"
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        return self._run(x)
